+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Binding topic to a specific output via `ulog_topic_add(TOPIC, OUTPUT, ENABLED)`
 - New status code: `ULOG_STATUS_NOT_FOUND` returned when a topic or output is not present (previously returned `ULOG_STATUS_ERROR`)
 - `ulog_topic_remove(TOPIC)`
+- `ulog_cleanup()` to free all dynamic resources and reset added entities (topics, outputs, etc.)
 
 ### Changed